requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.8.0
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.9.1
bleach==3.2.1
blinker==1.4
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
dominate==2.5.1
email-validator==1.1.1
entrypoints==0.3
feedparser==5.2.1
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-Script
Flask-Migrate
Flask-WTF==0.14.3
fsspec==0.8.5
idna==2.10
investpy==1.0
ipykernel==5.4.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
notebook-as-pdf==0.3.1
numpy==1.19.1
packaging==20.8
pandas==1.1.0
pandocfilters==1.4.3
parso==0.7.1
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
polygon-api-client==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
pyee==7.0.4
Pygments==2.7.3
pyinvesting==0.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.2
pyquery==1.4.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.24.0
retrying==1.3.3
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.5.0
sentry-sdk==0.19.5
six==1.15.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.18
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
Unidecode==1.1.2
urllib3==1.25.10
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
Yahoo-ticker-downloader==3.0.1
yfinance==0.1.55
mysqlclient
Flask-SocketIO
